PI3, all data, all tags
Upgraded the EB script to use the PI3 client along with the group tags utility to make sure all of the demands tags come as complete groups. The Pi3 client now also pulls data for the last 2 years each time so corrected historical data will be accounted for. This data request takes about 30mins with all of the chunks/retries etc.. That is still a marked improvement over the 5 hours that it used to take.

The areas document was modified to have an area value for each commodity which is necessary for certain building complexes that share meters.

Abandoned the kbtu pickle and csv. May be interesting to save a local copy of the data just incase the analysis needs to be tinkered with. This would safeguard the user from having to repeat the long pi data pull.